+++ CHANGELOG.md
@@ -41,6 +41,8 @@
 
 * Add the new "arDialect" method for Arabic dialect identification (i.e., Egyptian, Levantine, Maghrebi, and Peninsular) of the text (e.g., comments, reviews, etc.).
 
+* Add an Arabic version of the PHP [similar_text()](https://www.php.net/manual/en/function.similar-text.php) function, implemented using the [Needleman-Wunsch algorithm](https://en.wikipedia.org/wiki/Needleman-Wunsch_algorithm) with weighted scoring matrices and a non-linear gap penalty.
+
 [Top](#arphp-library---change-log)
 -->